fix for mail 2.8.1
