+++ lib/mail-iso-2022-jp/mail_under_280/common_methods_for_field.rb
