+++ lib/mail-iso-2022-jp/mail_under_280/invalid_encoding_error.rb
