+++ lib/mail-iso-2022-jp/mail_under_280/structured_fields.rb
@@ -1,0 +1,41 @@
+# coding: utf-8
+
+require 'mail'
+
+module Mail
+  class FromField < StructuredField
+    include FieldWithIso2022JpEncoding
+  end
+
+  class SenderField < StructuredField
+    include FieldWithIso2022JpEncoding
+  end
+
+  class ToField < StructuredField
+    include FieldWithIso2022JpEncoding
+  end
+
+  class CcField < StructuredField
+    include FieldWithIso2022JpEncoding
+  end
+
+  class ReplyToField < StructuredField
+    include FieldWithIso2022JpEncoding
+  end
+
+  class ResentFromField < StructuredField
+    include FieldWithIso2022JpEncoding
+  end
+
+  class ResentSenderField < StructuredField
+    include FieldWithIso2022JpEncoding
+  end
+
+  class ResentToField < StructuredField
+    include FieldWithIso2022JpEncoding
+  end
+
+  class ResentCcField < StructuredField
+    include FieldWithIso2022JpEncoding
+  end
+end
