+++ lib/mail-iso-2022-jp/ruby18/field_with_iso_2022_jp_encoding.rb
