+++ lib/mail-iso-2022-jp/ruby18/mail.rb
