+++ lib/mail-iso-2022-jp/structured_fields.rb
@@ -3,39 +3,9 @@
 require 'mail'
 
 module Mail
-  class FromField < StructuredField
-    include FieldWithIso2022JpEncoding
-  end
-
-  class SenderField < StructuredField
-    include FieldWithIso2022JpEncoding
-  end
-
-  class ToField < StructuredField
-    include FieldWithIso2022JpEncoding
-  end
-
-  class CcField < StructuredField
-    include FieldWithIso2022JpEncoding
-  end
-
-  class ReplyToField < StructuredField
-    include FieldWithIso2022JpEncoding
-  end
-
-  class ResentFromField < StructuredField
-    include FieldWithIso2022JpEncoding
-  end
-
-  class ResentSenderField < StructuredField
-    include FieldWithIso2022JpEncoding
-  end
-
-  class ResentToField < StructuredField
-    include FieldWithIso2022JpEncoding
-  end
-
-  class ResentCcField < StructuredField
-    include FieldWithIso2022JpEncoding
+  [FromField, SenderField, ToField, CcField, ReplyToField, ResentFromField, ResentSenderField, ResentToField, ResentCcField].each do |field_class|
+    field_class.class_eval do
+      include FieldWithIso2022JpEncoding
+    end
   end
 end
